refactor(gen): split static data out of docs.ts
- Move CODE_SAMPLES + RESPONSE_EXAMPLES to docs-samples.ts
- Move the CSS strings (header, Scalar overrides, guides) to docs-styles.ts
- docs.ts drops 1596 -> 821 lines; pure move, no behavior change
- Logic, schema-merge, markdown assembly, and routes stay in docs.ts
