+++ gen.pollinations.ai/src/routes/docs.ts
@@ -11,6 +11,14 @@ import { marked } from "marked";
 import { stringify as yamlStringify } from "yaml";
 import type { Env } from "@/env.ts";
 import LOGO_WHITE_SVG from "../../../assets/logo-text-white.svg?raw";
+import { CODE_SAMPLES, RESPONSE_EXAMPLES } from "./docs-samples.ts";
+import {
+    API_REFERENCE_CUSTOM_CSS,
+    GUIDES_CSS,
+    POLLINATIONS_HEADER_CSS,
+    POLLINATIONS_HEADER_SCALAR_CSS,
+    POLLINATIONS_HEADER_STANDALONE_CSS,
+} from "./docs-styles.ts";
 
 // Same favicon as enter.pollinations.ai (32×32 PNG, ~1.3kB → inlined as a
 // data URI so we don't need a separate binary asset route or a build-time
@@ -223,75 +231,6 @@ const LLM_DOC_SECTIONS: Record<string, string> = {
     mcp: MCP_SECTION,
 };
 
-const POLLINATIONS_HEADER_CSS = `
-.ph-bar {
-    position: fixed; top: 0; left: 0; right: 0; z-index: 2147483000;
-    min-height: 48px; box-sizing: border-box;
-    background: #0a0a0a; border-bottom: 1px solid #1f1f23;
-    display: flex; align-items: center; gap: 10px; padding: 8px 14px;
-    font: 13px/1.2 -apple-system, BlinkMacSystemFont, "Segoe UI", system-ui, sans-serif;
-    color: #d4d4d8;
-}
-.ph-bar .ph-brand {
-    color: #fafafa; text-decoration: none;
-    display: inline-flex; align-items: center; flex-shrink: 0;
-}
-.ph-bar .ph-brand img { height: 18px; width: auto; display: block; }
-@media (min-width: 640px) { .ph-bar .ph-brand img { height: 20px; } }
-@media (max-width: 640px) {
-    .ph-bar { justify-content: center; padding: 8px 12px; }
-    .ph-bar .ph-brand img { height: 22px; }
-}
-
-.ph-fab-cluster {
-    position: fixed; top: 64px; right: 18px; z-index: 9999;
-    display: flex; gap: 8px; align-items: center;
-    justify-content: flex-end; max-width: calc(100vw - 36px);
-}
-.ph-fab {
-    padding: 10px 14px; border-radius: 999px;
-    border: 1px solid #f59e0b; background: #f59e0b; color: #451a03;
-    font: 600 14px/1 -apple-system, BlinkMacSystemFont, "Segoe UI", system-ui, sans-serif;
-    box-shadow: 0 6px 20px rgba(245,158,11,.3);
-    cursor: pointer; text-decoration: none;
-    transition: background .15s, color .15s, box-shadow .15s, transform .1s;
-    display: inline-flex; align-items: center; gap: 6px; line-height: 1;
-}
-.ph-fab:hover {
-    background: #0a0a0a; color: #f59e0b;
-    box-shadow: 0 6px 24px rgba(245,158,11,.4);
-}
-.ph-fab:active { transform: translateY(1px); }
-.ph-fab svg { width: 16px; height: 16px; }
-@media (max-width: 640px) {
-    .ph-fab-cluster { top: 60px; right: 12px; }
-    .ph-fab { padding: 7px 11px; font-size: 12px; }
-    .ph-fab svg { width: 14px; height: 14px; }
-}
-`;
-
-const POLLINATIONS_HEADER_SCALAR_CSS = `
-/* Push Scalar's mount point down so its content (h1, version badges,
-   sidebar, mobile hamburger row) doesn't render under our fixed bar.
-   Body padding works because .ph-bar is position: fixed — it ignores
-   parent padding, so the bar stays at top while everything else shifts. */
-body { padding-top: 48px; }
-/* Match Scalar's mobile hamburger row background to our bar so the seam
-   between the two reads as one continuous header instead of a stripe. */
-@media (max-width: 1000px) {
-    .scalar-app [class*="lg:hidden"][class*="grid-area:header"] {
-        background: #0a0a0a !important;
-        border: 0 !important;
-        box-shadow: none !important;
-        backdrop-filter: none !important;
-    }
-}
-`;
-
-const POLLINATIONS_HEADER_STANDALONE_CSS = `
-body { padding-top: 48px; }
-`;
-
 type GuideId = "byop" | "cli" | "mcp";
 
 function pollinationsHeaderHtml(scalarHosted = false): string {
@@ -339,98 +278,6 @@ function pollinationsHeaderHtml(scalarHosted = false): string {
 })();
 </script>`;
 }
-
-const API_REFERENCE_CUSTOM_CSS = `
-/* Each table gets its own horizontal scroll, scoped to itself, so the
-   markdown container doesn't grow a second outer scrollbar that fights
-   with the scroll on adjacent code blocks. */
-.scalar-app .markdown table {
-  display: block;
-  max-width: 100%;
-  width: max-content;
-  overflow-x: auto;
-  -webkit-overflow-scrolling: touch;
-}
-.scalar-app .markdown table th,
-.scalar-app .markdown table td {
-  word-break: normal !important;
-  overflow-wrap: normal !important;
-}
-
-.scalar-app .markdown a {
-  color: #f59e0b;
-}
-.scalar-app .markdown a:hover {
-  color: #fbbf24;
-}
-
-/* Hide Scalar's native download UI — we surface it via the floating
-   action cluster (see .ph-fab-cluster below) for layout consistency. */
-.scalar-app .download-container { display: none !important; }
-
-/* Hide "Powered by Scalar" sidebar footer link. */
-.scalar-app a[href="https://www.scalar.com"] { display: none !important; }
-
-/* Hide Scalar's IDE/MCP quick-launch buttons (VS Code, Cursor, Generate MCP).
-   These render in the sidebar regardless of showDeveloperTools. We target
-   by URL scheme so the rule survives class-name renames across Scalar
-   versions, plus the section wrapper via :has() in case the buttons render
-   inside a labeled group. */
-.scalar-app a[href^="vscode:"],
-.scalar-app a[href^="vscode-insiders:"],
-.scalar-app a[href^="cursor:"],
-.scalar-app a[href*="mcp.scalar.com"],
-.scalar-app a[href*="generate-mcp"],
-.scalar-app a[href*="modelcontextprotocol"] { display: none !important; }
-.scalar-app section:has(> a[href^="vscode:"]),
-.scalar-app section:has(> a[href^="cursor:"]) { display: none !important; }
-
-/* Hide Scalar's "Ask AI" feature (sidebar button + any floating widget).
-   Targeted by attribute and class fragments, case-insensitive, since the
-   CDN bundle may rename the underlying classes between releases. */
-.scalar-app [class*="ask-ai" i],
-.scalar-app [class*="askai" i],
-.scalar-app [class*="ai-assistant" i],
-.scalar-app [aria-label*="Ask AI" i],
-.scalar-app [title*="Ask AI" i],
-.scalar-app button[data-feature="ask-ai" i] { display: none !important; }
-
-/* Full-width prose for sections with no right-column code samples (Quick Start,
-   Auth, BYOP, CLI, MCP, Errors, Safety, plain Models/Account). Scalar lays
-   each tag out as two flex columns; the right column stays empty for prose-
-   only tags, wasting ~50% of the page. We collapse the empty column and let
-   the prose stretch. The same rule fixes the section header row above. */
-.scalar-app .section-columns:has(> .section-column:nth-child(2):empty) > .section-column:first-child {
-    flex: 1 1 100% !important;
-    max-width: 100% !important;
-}
-.scalar-app .section-columns > .section-column:nth-child(2):empty {
-    display: none !important;
-}
-.scalar-app .section-header-wrapper:not(:has(> :nth-child(2))) {
-    grid-template-columns: 1fr !important;
-}
-
-/* "Show more" buttons (tagged by our header script — see scanScalarButtons).
-   Promote to the amber action color so they don't disappear into the page. */
-.scalar-app .ph-show-more {
-    background: #f59e0b !important;
-    color: #451a03 !important;
-    border-color: #f59e0b !important;
-}
-.scalar-app .ph-show-more:hover {
-    background: #fbbf24 !important;
-    border-color: #fbbf24 !important;
-}
-/* Force the chevron/triangle SVG inside the button to match text color
-   (otherwise it inherits Scalar's dim color and disappears on amber). */
-.scalar-app .ph-show-more svg,
-.scalar-app .ph-show-more svg * {
-    color: #451a03 !important;
-    fill: currentColor !important;
-    stroke: currentColor !important;
-}
-`;
 
 function generationDocumentation(): OpenApiSchema {
     return {
@@ -695,572 +542,6 @@ function isGenerationPath(path: string): boolean {
     );
 }
 
-// ---------------------------------------------------------------------------
-// x-codeSamples: multi-language examples injected into the OpenAPI schema
-// ---------------------------------------------------------------------------
-const CODE_SAMPLES: Record<
-    string,
-    { label: string; lang: string; source: string }[]
-> = {
-    "post /account/keys": [
-        {
-            label: "Create app key",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/account/keys \\
-  -H "Authorization: Bearer YOUR_SECRET_KEY" \\
-  -H "Content-Type: application/json" \\
-  -d '{
-    "name": "myapp",
-    "type": "publishable",
-    "redirectUris": ["https://myapp.com/callback"]
-  }'`,
-        },
-    ],
-    "post /v1/chat/completions": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/v1/chat/completions \\
-  -H "Authorization: Bearer YOUR_API_KEY" \\
-  -H "Content-Type: application/json" \\
-  -d '{
-    "model": "openai",
-    "messages": [{"role": "user", "content": "Hello!"}]
-  }'`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `from openai import OpenAI
-
-client = OpenAI(
-    base_url="https://gen.pollinations.ai/v1",
-    api_key="YOUR_API_KEY"
-)
-
-response = client.chat.completions.create(
-    model="openai",
-    messages=[{"role": "user", "content": "Hello!"}]
-)
-print(response.choices[0].message.content)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `import OpenAI from "openai";
-
-const client = new OpenAI({
-  baseURL: "https://gen.pollinations.ai/v1",
-  apiKey: "YOUR_API_KEY",
-});
-
-const response = await client.chat.completions.create({
-  model: "openai",
-  messages: [{ role: "user", content: "Hello!" }],
-});
-console.log(response.choices[0].message.content);`,
-        },
-    ],
-    "get /text/{prompt}": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl "https://gen.pollinations.ai/text/Write%20a%20haiku?model=openai" \\
-  -H "Authorization: Bearer YOUR_API_KEY"`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/text/Write a haiku",
-    params={"model": "openai"},
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-print(response.text)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/text/Write%20a%20haiku?model=openai",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-console.log(await response.text());`,
-        },
-    ],
-    "get /image/{prompt}": [
-        {
-            label: "HTML",
-            lang: "HTML",
-            source: `<!-- No code needed — use as an image URL -->
-<img src="https://gen.pollinations.ai/image/a%20cat%20in%20space?model=flux" />`,
-        },
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `# Generate an image
-curl "https://gen.pollinations.ai/image/a%20cat%20in%20space?model=flux" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o image.jpg
-
-# Generate a video
-curl "https://gen.pollinations.ai/image/a%20sunset%20timelapse?model=veo&duration=4" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o video.mp4`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/image/a cat in space",
-    params={"model": "flux"},
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-with open("image.jpg", "wb") as f:
-    f.write(response.content)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/image/a%20cat%20in%20space?model=flux",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const blob = await response.blob();`,
-        },
-    ],
-    "get /video/{prompt}": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl "https://gen.pollinations.ai/video/a%20sunset%20timelapse?model=veo&duration=4" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o video.mp4`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/video/a sunset timelapse",
-    params={"model": "veo", "duration": 4},
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-with open("video.mp4", "wb") as f:
-    f.write(response.content)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/video/a%20sunset%20timelapse?model=veo&duration=4",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const blob = await response.blob();`,
-        },
-    ],
-    "get /audio/{text}": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `# Text-to-speech
-curl "https://gen.pollinations.ai/audio/Hello%20world?voice=nova" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o speech.mp3
-
-# Generate music (ElevenLabs)
-curl "https://gen.pollinations.ai/audio/upbeat%20jazz?model=elevenmusic&duration=30" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o music.mp3
-
-# Generate music (ACE-Step, open-source)
-curl "https://gen.pollinations.ai/audio/brazilian%20berimbau%20instrumental?model=acestep&duration=15" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o music.mp3`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/audio/Hello world",
-    params={"voice": "nova"},
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-with open("speech.mp3", "wb") as f:
-    f.write(response.content)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/audio/Hello%20world?voice=nova",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const audio = await response.blob();`,
-        },
-    ],
-    "post /v1/audio/speech": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/v1/audio/speech \\
-  -H "Authorization: Bearer YOUR_API_KEY" \\
-  -H "Content-Type: application/json" \\
-  -d '{"input": "Hello world", "voice": "nova"}' \\
-  -o speech.mp3`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `from openai import OpenAI
-
-client = OpenAI(
-    base_url="https://gen.pollinations.ai/v1",
-    api_key="YOUR_API_KEY",
-)
-
-response = client.audio.speech.create(
-    model="tts-1",
-    voice="nova",
-    input="Hello world",
-)
-response.stream_to_file("speech.mp3")`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `import OpenAI from "openai";
-
-const client = new OpenAI({
-  baseURL: "https://gen.pollinations.ai/v1",
-  apiKey: "YOUR_API_KEY",
-});
-
-const response = await client.audio.speech.create({
-  model: "tts-1",
-  voice: "nova",
-  input: "Hello world",
-});
-const buffer = Buffer.from(await response.arrayBuffer());`,
-        },
-    ],
-    "post /v1/audio/transcriptions": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/v1/audio/transcriptions \\
-  -H "Authorization: Bearer YOUR_API_KEY" \\
-  -F file=@audio.mp3 \\
-  -F model=whisper-large-v3`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `from openai import OpenAI
-
-client = OpenAI(
-    base_url="https://gen.pollinations.ai/v1",
-    api_key="YOUR_API_KEY",
-)
-
-with open("audio.mp3", "rb") as f:
-    transcript = client.audio.transcriptions.create(
-        model="whisper-large-v3", file=f
-    )
-print(transcript.text)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `import OpenAI from "openai";
-import fs from "fs";
-
-const client = new OpenAI({
-  baseURL: "https://gen.pollinations.ai/v1",
-  apiKey: "YOUR_API_KEY",
-});
-
-const transcript = await client.audio.transcriptions.create({
-  model: "whisper-large-v3",
-  file: fs.createReadStream("audio.mp3"),
-});
-console.log(transcript.text);`,
-        },
-    ],
-    "get /account/balance": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/account/balance \\
-  -H "Authorization: Bearer YOUR_API_KEY"`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/account/balance",
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-print(response.json())  # {"balance": 42.5}`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/account/balance",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const { balance } = await response.json();
-console.log(balance);`,
-        },
-    ],
-    "get /account/profile": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/account/profile \\
-  -H "Authorization: Bearer YOUR_API_KEY"`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/account/profile",
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-profile = response.json()
-print(profile["githubUsername"])`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/account/profile",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const profile = await response.json();
-console.log(profile.githubUsername);`,
-        },
-    ],
-    "get /account/key": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/account/key \\
-  -H "Authorization: Bearer YOUR_API_KEY"`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/account/key",
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-key_info = response.json()
-print(f"Valid: {key_info['valid']}, Type: {key_info['type']}")`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch(
-  "https://gen.pollinations.ai/account/key",
-  { headers: { Authorization: "Bearer YOUR_API_KEY" } },
-);
-const keyInfo = await response.json();
-console.log(\`Valid: \${keyInfo.valid}, Type: \${keyInfo.type}\`);`,
-        },
-    ],
-    "get /v1/models": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/v1/models`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `from openai import OpenAI
-
-client = OpenAI(
-    base_url="https://gen.pollinations.ai/v1",
-    api_key="YOUR_API_KEY",
-)
-
-models = client.models.list()
-for model in models.data:
-    print(model.id)`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `import OpenAI from "openai";
-
-const client = new OpenAI({
-  baseURL: "https://gen.pollinations.ai/v1",
-  apiKey: "YOUR_API_KEY",
-});
-
-const models = await client.models.list();
-models.data.forEach((m) => console.log(m.id));`,
-        },
-    ],
-    "get /image/models": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/image/models`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch("https://gen.pollinations.ai/image/models");
-const models = await response.json();
-models.forEach((m) => console.log(\`\${m.id}: \${m.description}\`));`,
-        },
-    ],
-    "get /text/models": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/text/models`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch("https://gen.pollinations.ai/text/models");
-const models = await response.json();
-models.forEach((m) => console.log(\`\${m.id}: \${m.description}\`));`,
-        },
-    ],
-    "get /audio/models": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl https://gen.pollinations.ai/audio/models`,
-        },
-        {
-            label: "JavaScript",
-            lang: "JavaScript",
-            source: `const response = await fetch("https://gen.pollinations.ai/audio/models");
-const models = await response.json();
-models.forEach((m) => console.log(\`\${m.id}: \${m.description}\`));`,
-        },
-    ],
-    "get /account/usage": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `# Get usage history (JSON)
-curl "https://gen.pollinations.ai/account/usage?limit=10" \\
-  -H "Authorization: Bearer YOUR_API_KEY"
-
-# Export the latest 50,000 rows from the last 30 days as CSV
-curl "https://gen.pollinations.ai/account/usage?format=csv&days=30&limit=50000" \\
-  -H "Authorization: Bearer YOUR_API_KEY" -o usage.csv`,
-        },
-        {
-            label: "Python",
-            lang: "Python",
-            source: `import requests
-
-response = requests.get(
-    "https://gen.pollinations.ai/account/usage",
-    params={"limit": 10, "days": 30},
-    headers={"Authorization": "Bearer YOUR_API_KEY"},
-)
-for record in response.json()["usage"]:
-    print(f"{record['model']}: {record['cost_usd']} pollen")`,
-        },
-    ],
-    "get /account/usage/daily": [
-        {
-            label: "cURL",
-            lang: "Shell",
-            source: `curl "https://gen.pollinations.ai/account/usage/daily?days=30" \\
-  -H "Authorization: Bearer YOUR_API_KEY"`,
-        },
-    ],
-};
-
-// ---------------------------------------------------------------------------
-// Response examples injected into the OpenAPI schema
-// ---------------------------------------------------------------------------
-const RESPONSE_EXAMPLES: Record<string, unknown> = {
-    "post /v1/chat/completions": {
-        id: "chatcmpl-abc123",
-        object: "chat.completion",
-        created: 1700000000,
-        model: "openai",
-        choices: [
-            {
-                index: 0,
-                message: {
-                    role: "assistant",
-                    content: "Hello! How can I help you today?",
-                },
-                finish_reason: "stop",
-            },
-        ],
-        usage: {
-            prompt_tokens: 10,
-            completion_tokens: 12,
-            total_tokens: 22,
-        },
-    },
-    "get /v1/models": {
-        object: "list",
-        data: [
-            {
-                id: "openai",
-                object: "model",
-                created: 1700000000,
-                owned_by: "pollinations",
-            },
-            {
-                id: "claude",
-                object: "model",
-                created: 1700000000,
-                owned_by: "pollinations",
-            },
-            {
-                id: "gemini",
-                object: "model",
-                created: 1700000000,
-                owned_by: "pollinations",
-            },
-        ],
-    },
-    "get /account/balance": {
-        balance: 42.5,
-    },
-    "get /account/profile": {
-        githubUsername: "janedeveloper",
-        image: "https://avatars.example.com/jane.jpg",
-        name: "Jane Developer",
-        email: "jane@example.com",
-    },
-    "get /account/key": {
-        valid: true,
-        type: "secret",
-        name: "my-bot",
-        expiresAt: null,
-        expiresIn: null,
-        permissions: {
-            models: null,
-            account: ["usage"],
-        },
-        pollenBudget: null,
-        rateLimitEnabled: false,
-    },
-};
-
 function injectSamples(schema: OpenApiSchema): OpenApiSchema {
     const paths = schema.paths as Record<string, unknown> | undefined;
     if (!paths) return schema;
@@ -1422,62 +703,6 @@ const GUIDES: Guide[] = [
 ];
 
 const GUIDES_BY_ID = new Map(GUIDES.map((g) => [g.id, g]));
-
-const GUIDES_CSS = `
-:root { color-scheme: dark; }
-* { box-sizing: border-box; }
-html, body { overflow-x: hidden; }
-body {
-    font: 15px/1.65 -apple-system, BlinkMacSystemFont, "Segoe UI", system-ui, sans-serif;
-    color: #d4d4d8;
-    background: #0a0a0a;
-    margin: 0;
-}
-.wrap { max-width: 800px; margin: 0 auto; padding: 1.5rem 1.25rem 4rem; }
-h1, h2, h3, h4 { font-weight: 600; line-height: 1.3; margin: 2rem 0 .75rem; color: #fafafa; }
-h1 { font-size: 1.875rem; margin-top: 1rem; }
-h2 { font-size: 1.375rem; padding-bottom: .375rem; border-bottom: 1px solid #27272a; }
-h3 { font-size: 1.125rem; }
-p, li, blockquote { overflow-wrap: anywhere; word-break: break-word; }
-p { margin: .75rem 0; }
-code {
-    background: #18181b; padding: 2px 5px; border-radius: 4px; color: #f4f4f5;
-    font: 13px ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
-    overflow-wrap: anywhere; word-break: break-word;
-}
-pre {
-    background: #000; color: #f1f5f9; padding: 1rem; border-radius: 8px;
-    border: 1px solid #1f1f23; overflow-x: auto; font-size: 13px; line-height: 1.5;
-    max-width: 100%;
-}
-pre code {
-    background: transparent; color: inherit; padding: 0;
-    overflow-wrap: normal; word-break: normal; white-space: pre;
-}
-table {
-    display: block; max-width: 100%; width: max-content;
-    overflow-x: auto; -webkit-overflow-scrolling: touch;
-    border-collapse: collapse; margin: 1rem 0; font-size: .9rem;
-}
-th, td { padding: .5rem .75rem; border: 1px solid #27272a; text-align: left; vertical-align: top; }
-th { background: #18181b; font-weight: 600; color: #fafafa; }
-a { color: #f59e0b; text-decoration: none; overflow-wrap: anywhere; word-break: break-word; }
-a:hover { text-decoration: underline; }
-blockquote { border-left: 3px solid #3f3f46; padding: .25rem 1rem; color: #a1a1aa; margin: 1rem 0; background: #131316; }
-ul, ol { padding-left: 1.5rem; }
-li { margin: .25rem 0; }
-hr { border: 0; border-top: 1px solid #27272a; margin: 2rem 0; }
-img { max-width: 100%; height: auto; }
-.guide-cards { display: grid; gap: 1rem; grid-template-columns: 1fr; margin-top: 1.5rem; }
-@media (min-width: 600px) { .guide-cards { grid-template-columns: 1fr 1fr; } }
-.guide-card {
-    display: block; padding: 1.25rem; border: 1px solid #27272a; border-radius: 10px;
-    background: #131316; color: #fafafa; text-decoration: none; transition: border-color .15s, background .15s;
-}
-.guide-card:hover { border-color: #f59e0b; background: #161618; }
-.guide-card h3 { margin: 0 0 .375rem; font-size: 1.05rem; color: #fafafa; }
-.guide-card p { margin: 0; color: #a1a1aa; font-size: .9rem; }
-`;
 
 function guidesPage(body: string): string {
     return `<!doctype html>
